Moved all classes in Data.cs to their own file.
